Refactor layout builder source code
Modified the `getSectionType` function in LayoutBase.php to include type hinting for the $paragraphType parameter and adjusted its return type to explicitly be a string. Also, adjusted 'remove_gutters' value from '1' to '0' and refined the block parameters in the comments. Removed an unused line of code pertaining to block content loading.