Update version in osu_migrations.info.yml
The version number in osu_migrations.info.yml file has been updated from 1.0.17 to 1.0.18. This increment in version number represents changes in content migration labels and adjustment of dependencies within the project. A README file has also been introduced to provide details on further updates and views that need to be rebuilt.